requirements.txt
torch
transformers
Pillow
fastapi
uvicorn
python-multipart